refactor: extract StackBarGraph into its own component
Move the stacked bar out of TableSummaryReport so it can be composed as a footer or used standalone, and export it from the public API.
