+++ src/components/TableSummaryReport/TableSummaryReport.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent } from "@testing-library/react";
-import type { StackBarGraphSegment, TableSummaryReportProps } from "src/components/TableSummaryReport";
-import { StackBarGraph, TableSummaryReport } from "src/components/TableSummaryReport";
+import { StackBarGraph, type StackBarGraphSegment } from "src/components/StackBarGraph";
+import { TableSummaryReport, type TableSummaryReportProps } from "src/components/TableSummaryReport";
 import { click, render } from "src/utils/rtl";
 import { vi } from "vitest";
 
@@ -82,54 +82,8 @@ describe("TableSummaryReport", () => {
     expect(onMetricClick).toHaveBeenCalledWith("missing");
   });
 
-  it("shows zero percent for segments with a zero total", async () => {
-    // Given stack bar segments with no items
-    const r = await render(
-      <TableSummaryReport
-        {...createProps({
-          footer: (
-            <StackBarGraph
-              title="Coverage by status"
-              totalLabel="Cost Codes"
-              segments={[
-                { label: "Complete", count: 0, status: "success" },
-                { label: "Missing", count: 0, status: "error" },
-              ]}
-            />
-          ),
-        })}
-      />,
-    );
-    // Then the legend does not render invalid percentages
-    expect(r.stackBarGraph_legend).toHaveTextContent("0% Complete (0)");
-    expect(r.stackBarGraph_legend).toHaveTextContent("0% Missing (0)");
-  });
-
-  it("sizes stack bar segments according to their counts", async () => {
-    // Given a stack bar with unequal segment counts
-    const r = await render(
-      <TableSummaryReport
-        {...createProps({
-          footer: (
-            <StackBarGraph
-              title="Coverage by status"
-              totalLabel="Cost Codes"
-              segments={[
-                { label: "Complete", count: 3, status: "success" },
-                { label: "Missing", count: 1, status: "error" },
-              ]}
-            />
-          ),
-        })}
-      />,
-    );
-    // Then the bar uses counts as flex proportions
-    expect(r.stackBarGraph_bar.children[0]).toHaveStyle({ flexGrow: "3" });
-    expect(r.stackBarGraph_bar.children[1]).toHaveStyle({ flexGrow: "1" });
-  });
-
-  it("renders a caller-provided stack bar title", async () => {
-    // Given a custom stack bar title
+  it("renders a caller-provided footer", async () => {
+    // Given a custom stack bar in the footer
     const r = await render(
       <TableSummaryReport
         {...createProps({
@@ -137,7 +91,7 @@ describe("TableSummaryReport", () => {
         })}
       />,
     );
-    // Then the custom title is shown
+    // Then the composed footer is shown
     expect(r.stackBarGraph).toHaveTextContent("Bid package status");
   });
 });
@@ -149,9 +103,7 @@ function createProps(overrides: Partial<TableSummaryReportProps<string>> = {}): 
     metrics: [{ value: "missing", label: "Missing", count: 4, status: "error" }],
     issueLabel: "View Issues",
     onIssueClick: () => {},
-    footer: footer ?? (
-      <StackBarGraph title="Coverage by status" totalLabel="Cost Codes" segments={defaultSegments()} />
-    ),
+    footer: footer ?? <StackBarGraph title="Coverage by status" totalLabel="Cost Codes" segments={defaultSegments()} />,
     ...rest,
   };
 }
