+++ src/components/TableSummaryReport/index.ts
@@ -1,8 +1,2 @@
-export { StackBarGraph, TableSummaryReport } from "./TableSummaryReport";
-export type {
-  StackBarGraphProps,
-  StackBarGraphSegment,
-  TableSummaryReportMetric,
-  TableSummaryReportProps,
-  TableSummaryReportStatus,
-} from "./TableSummaryReport";
+export { TableSummaryReport } from "./TableSummaryReport";
+export type { TableSummaryReportMetric, TableSummaryReportProps, TableSummaryReportStatus } from "./TableSummaryReport";
